Move HRN resolution into payment module
Keep human-readable-name resolution beside the unified payment code
that consumes it. This commit only relocates the existing resolver and
updates its internal import paths.
